Fixed Anim.c16 bug, memory index was signed
git-svn-id: https://mash16.googlecode.com/svn/trunk@57 f32040c0-604e-ac78-71ca-de615562ae79

+++ src/core/cpu.c
@@ -245,7 +245,7 @@ void op_drw_r(cpu_state* state)
 {
     int16_t x = state->r[state->i.yx & 0x0f];
     int16_t y = state->r[state->i.yx >> 4];
-    state->f.c = op_drw(&state->m[state->r[state->i.z]],
+    state->f.c = op_drw(&state->m[(uint16_t)state->r[state->i.z]],
         state->vm, x, y, state->sw, state->sh,
         state->fx, state->fy);
 }
